+++ src/main/java/org/liftoff/project/models/dto/LoginDTO.java
@@ -7,32 +7,32 @@ import java.util.Date;
 
 public class LoginDTO {
     @NotNull
-    @NotBlank
+    @NotBlank(message = "Username is required!")
     @Size(min = 5, max = 30, message = "Invalid username. Must be between 5 and 30 characters.")
     private String username;
 
     @NotNull
-    @NotBlank
+    @NotBlank(message = "Password is required!")
     @Size(min = 5, max = 30, message = "Invalid password. Must be between 5 and 30 characters.")
     private String password;
 
 
-    @NotNull
-    @NotBlank
-    @Size( max = 4, message = "Invalid date of birth Must be 4 characters.")
-    private Date dateOfBirth;
+//    @NotNull
+//    @NotBlank
+//    @Size( max = 4, message = "Invalid date of birth Must be 4 characters.")
+//    private Date dateOfBirth;
 
 
-    @NotNull
-    @NotBlank
-    @Size(min = 5, max = 30, message = "Invalid email. Must be between 5 and 30 characters.")
-    private String email;
+//    @NotNull
+//    @NotBlank
+//    @Size(min = 5, max = 30, message = "Invalid email. Must be between 5 and 30 characters.")
+//    private String email;
 
 
-    @NotNull
-    @NotBlank
-    @Size(min = 5, max = 30, message = "Invalid password. Must be between 5 and 30 characters.")
-    private Integer phoneNumber;
+//    @NotNull
+//    @NotBlank
+//    @Size(min = 5, max = 30, message = "Invalid password. Must be between 5 and 30 characters.")
+//    private Integer phoneNumber;
 
     public String getUsername() {
 
@@ -54,32 +54,32 @@ public class LoginDTO {
         this.password = password;
     }
 
-   public Date getDateOfBirth() {
-
-        return dateOfBirth;
+//   public Date getDateOfBirth() {
+//
+//        return dateOfBirth;
+//    }
+//
+//    public void setDateOfBirth(Date dateOfBirth) {
+//        this.dateOfBirth = dateOfBirth;
+//    }
+//    public String getEmail() {
+//
+//        return email;
+//   }
+//
+//    public void setEmail(String email) {
+//
+//        this.email = email;
     }
 
-    public void setDateOfBirth(Date dateOfBirth) {
-        this.dateOfBirth = dateOfBirth;
-    }
-    public String getEmail() {
-
-        return email;
-   }
-
-    public void setEmail(String email) {
-
-        this.email = email;
-    }
-
-    public Integer getPhoneNumber() {
-
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(Integer phoneNumber) {
-
-        this.phoneNumber = phoneNumber;
-   }
-
-}
+//    public Integer getPhoneNumber() {
+//
+//        return phoneNumber;
+//    }
+//
+//    public void setPhoneNumber(Integer phoneNumber) {
+//
+//        this.phoneNumber = phoneNumber;
+//   }
+//
+//}
