+++ src/main/resources/static/styles.css
@@ -52,3 +52,7 @@ a:link {
   box-shadow: rgba(0, 0, 0, .3) 2px 8px 4px -6px;
 
 }
+
+.error {
+color: red;
+}
